Refactor KV cache trace types into inference module
